move servers into modules
